Move common args
