+++ launch_pal/robot_arguments/__init__.py
@@ -16,7 +16,6 @@
 """robot_arguments Module."""
 
 from .ari import AriArgs
-from .tiago import TiagoArgs
 from .tiago_dual import TiagoDualArgs
 from .tiago_sea import TiagoSEAArgs
 from .tiago_sea_dual import TiagoSEADualArgs
@@ -24,15 +23,16 @@ from .tiago_pro import TiagoProArgs
 from .omni_base import OmniBaseArgs
 from .pmb2 import PMB2Args
 from .robot import RobotArgs
+from .common import CommonArgs
 
 __all__ = [
     'AriArgs',
-    'TiagoArgs',
     'TiagoDualArgs',
     'TiagoSEAArgs',
     'TiagoSEADualArgs',
     'TiagoProArgs',
     'OmniBaseArgs',
     'PMB2Args',
-    'RobotArgs'
+    'RobotArgs',
+    'CommonArgs'
 ]
